- changelog; - addressed PR comments
